src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/remove-and-insert-style-expected.txt

CONSOLE ERROR: line 48: Uncaught TypeError: host.createShadowRoot is not a function
Test for Issue: 246300: Styles in nested shadows are not recalculated correctly on insertion.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
